+++ server/component/style/link/LinkView.php
@@ -70,12 +70,13 @@ class LinkView extends StyleView
      * the data for the entry value 
      */
     public function output_content_entry($entry_value)
-    {
+    {        
         if($this->url == "") return;
+        $entry_data = $entry_value;
         $this->url = $this->model->get_entry_value($entry_value, $this->model->get_db_field("url"));
         if($this->label == "") $this->label = htmlspecialchars($this->url);
         $target = ($this->open_in_new_tab) ? 'target="_blank"' : "";
-        require __DIR__ . "/tpl_link.php";
+        require __DIR__ . "/tpl_link_entryValue.php";
     }
 
     /**
@@ -85,12 +86,13 @@ class LinkView extends StyleView
      */
     public function output_content_mobile_entry($entry_value)
     {
-        $style = parent::output_content_mobile();
+        $style = parent::output_content_mobile_entry($entry_value);
         $this->url = $this->model->get_db_field("url");
         $this->url = $this->get_entry_value($entry_value, $this->url); 
         $style['url']['content'] = $this->url[0] == '/' ? $this->url : '/' . $this->url;
         return $style;
     }
+
 	
 }
 ?>